refactor(controlplane): split k8s_pool.go into cohesive same-package files (#722 phase 2)
Behavior-preserving mechanical split of controlplane/k8s_pool.go (~3,482 lines)
into cohesive files under the same `kubernetes` build tag, the same treatment PR
#728 gave server/conn.go. Pure move: zero renames, zero signature changes, zero
logic edits.

New files (all carry the `//go:build kubernetes` constraint header):
- k8s_pool.go         — struct, constants, constructors, resolveCPUID,
                        workerServiceAccountName, ensureLifecycle, SetMaxWorkers,
                        SetWorkerResources, boolPtr/int64Ptr, interface check
- k8s_pool_spawn.go   — pod spec construction + spawn path (SpawnWorker,
                        spawnWorker, controlPlaneIDLabelValue, createPodWithBackoff,
                        podReadyInfo, waitForPodReady, waitForWorkerTCP[WithMetadata],
                        spawnWorkerBackground, spawnReservedWorker, workerResources*)
- k8s_pool_acquire.go — acquisition/scheduling/reserve (AcquireWorker, ReleaseWorker,
                        Worker, ActivateReservedWorker, ReserveSharedWorker,
                        reserveClaimedWorker, claimSpecificWorker, adoptClaimedWorker,
                        connectWorker*, checkReservedWorkerLiveness, SpawnMinWorkers,
                        node-seen + scheduling helpers)
- k8s_pool_lifecycle.go — retire/drain/health (RetireWorker*, RetireIfDrainingAndEmpty,
                        TransitionToHotIdleIfNoSessions, retireClaimedWorker,
                        DeleteWorkerArtifacts, deleteRetiredRuntimeWorker,
                        HealthCheckLoop, ShutdownAll, retireWorkerPod, idleReaper,
                        reapIdleWorkers, reapStuckActivatingWorkers,
                        cleanDeadWorkersLocked, lease snapshot/disposition helpers,
                        markWorkerRetired*Locked)
- k8s_pool_reconcile.go — janitor reconcilers + informer (RetireOneMismatchedVersionWorker,
                        workerImageForPod, cleanupOrphanedWorkerPods/Secrets,
                        startInformer, onPodTerminated)
- k8s_pool_helpers.go — small helpers (allocate*IDLocked, persistWorkerRecord,
                        workerRecordFor, healthCheckPayload*, podName*, workerPodName*,
                        trimK8sPodHashSuffix)

spawnWorker (DUCKGRES_DUCKDB_MAX_SESSIONS=1, terminationGracePeriodSeconds=3600),
AcquireWorker, and all session-counting / drain-token logic are moved verbatim,
not edited. No AcquireWorker de-dup (left for #722 phase 3).
